refactor(core): unify the Windows transient-lock retry into meridian-core
Follow-up to #606 and #607, addressing #607's 🟡 review note: both PRs
independently implemented "retry a fallible op with linear backoff to ride
out a transient Windows file lock" — #606 sync (std::fs) in db_crypto, #607
async (tokio::fs) in the tray's backend_install. Now that both have merged,
fold them into one shared home.

- New `meridian_core::retry` module with two variants of one schedule
  (base_delay × attempt backoff, last attempt's error surfaced):
  `retry_transient` (async, tokio::time) and `retry_transient_blocking`
  (sync, std::thread). Both pub, platform-neutral, unit-tested here (8 tests
  — the four boundary properties × both variants) so they run on macOS CI.
- `db_crypto::rename_with_retry` now delegates to
  `retry_transient_blocking`; the tray's `rename_with_retry` to
  `retry_transient`. Both drop their private copies of the loop.
- meridian-core's tokio gains the `time` feature (already built with `full`
  by every binary that links this crate, so per the existing note it adds
  nothing new to compile).

`wait_until_gone` (the tray's daemon-exit poll) stays local — it's a
poll-until-empty shape, not a retry-an-op, and Windows-specific. No behavior
change: same attempt counts, same backoff, same surfaced errors.
