+++ tray/src-tauri/src/backend_install.rs
@@ -742,45 +742,15 @@ async fn migrate_legacy_bundle_env(home: &Path) {
 const RENAME_ATTEMPTS: u32 = 5;
 const RENAME_BASE_DELAY: Duration = Duration::from_millis(100);
 
-/// Retry an async fallible operation up to `attempts` times with linear backoff
-/// (`base_delay` × the attempt number between tries), surfacing the last error
-/// only after every attempt fails. On Windows a swap over a binary another
-/// process momentarily holds — an antivirus scan, the Search Indexer — fails
-/// transiently with os error 32; a few spaced retries absorb that without
-/// masking a genuinely stuck file. Platform-neutral and generic over the
-/// operation so the retry logic is unit-tested on macOS CI even though the
-/// transient failures it guards are Windows-specific.
-async fn retry_transient<F, Fut, T, E>(
-    attempts: u32,
-    base_delay: Duration,
-    mut op: F,
-) -> Result<T, E>
-where
-    F: FnMut() -> Fut,
-    Fut: std::future::Future<Output = Result<T, E>>,
-{
-    let mut attempt = 1;
-    loop {
-        match op().await {
-            Ok(v) => return Ok(v),
-            Err(e) if attempt >= attempts => return Err(e),
-            Err(_) => {
-                tokio::time::sleep(base_delay * attempt).await;
-                attempt += 1;
-            }
-        }
-    }
-}
-
-/// `tokio::fs::rename` with the bounded transient-failure retry
-/// [`retry_transient`] provides. [`stage_binary`]'s rename swaps the freshly
-/// staged binary over `~/.meridian/bin/meridian(.exe)` — the live daemon's image
-/// path, exactly where a momentary Windows file hold lands — so retrying turns a
-/// self-clearing blip into a silent success instead of an install-failed notice.
-/// A persistent failure still returns the real error so a genuinely locked file
-/// is not swept under the rug.
+/// `tokio::fs::rename` with the bounded transient-failure retry from the shared
+/// [`meridian_core::retry::retry_transient`]. [`stage_binary`]'s rename swaps the
+/// freshly staged binary over `~/.meridian/bin/meridian(.exe)` — the live
+/// daemon's image path, exactly where a momentary Windows file hold lands — so
+/// retrying turns a self-clearing blip into a silent success instead of an
+/// install-failed notice. A persistent failure still returns the real error so a
+/// genuinely locked file is not swept under the rug.
 async fn rename_with_retry(from: &Path, to: &Path) -> std::io::Result<()> {
-    retry_transient(RENAME_ATTEMPTS, RENAME_BASE_DELAY, || {
+    meridian_core::retry::retry_transient(RENAME_ATTEMPTS, RENAME_BASE_DELAY, || {
         tokio::fs::rename(from, to)
     })
     .await
@@ -1228,97 +1198,19 @@ mod tests {
     }
 
     // ---------------------------------------------------------------------
-    // Windows install-lock hardening (the retry/poll primitives behind
-    // `rename_with_retry` and `stop_running_daemon_before_stage`).
+    // Windows install-lock hardening: `rename_with_retry` (the tray-side thin
+    // wrapper over the shared `meridian_core::retry::retry_transient`) and the
+    // daemon-exit poll `wait_until_gone`. The retry primitive itself is
+    // unit-tested in `meridian-core::retry`; these cover the wrapper's wiring
+    // and the poll loop.
     //
-    // The failures these guard against — os error 32 when overwriting a
+    // The failures they guard against — os error 32 when overwriting a
     // still-open `meridian.exe`, a daemon that needs a beat to release its
     // binary — only occur on Windows, but the *logic* is deliberately
     // platform-neutral so it compiles and these tests actually run on the
-    // macOS-only CI. They use zero-delay backoff and a call-counting fake so
+    // macOS-only CI. They use zero-delay backoff and call-counting fakes so
     // they are deterministic, fast, and free of any real timing dependence.
     // ---------------------------------------------------------------------
-
-    /// `retry_transient` returns a first-attempt success straight through and
-    /// never invokes the operation a second time — the common, fast path.
-    #[tokio::test]
-    async fn retry_transient_returns_on_first_success_without_retrying() {
-        use std::sync::atomic::{AtomicU32, Ordering};
-        let calls = AtomicU32::new(0);
-        let out: Result<u32, ()> = retry_transient(5, Duration::ZERO, || {
-            calls.fetch_add(1, Ordering::SeqCst);
-            async { Ok(7) }
-        })
-        .await;
-        assert_eq!(out, Ok(7));
-        assert_eq!(
-            calls.load(Ordering::SeqCst),
-            1,
-            "a first-attempt success must not retry"
-        );
-    }
-
-    /// `retry_transient` keeps retrying across transient failures and returns
-    /// the eventual success — the exact shape of an AV/indexer hold that lets
-    /// go after a moment. Verifies it recovers *and* stops as soon as it does.
-    #[tokio::test]
-    async fn retry_transient_recovers_after_transient_failures() {
-        use std::sync::atomic::{AtomicU32, Ordering};
-        let calls = AtomicU32::new(0);
-        let out: Result<&str, &str> = retry_transient(5, Duration::ZERO, || {
-            let n = calls.fetch_add(1, Ordering::SeqCst);
-            async move {
-                if n < 3 {
-                    Err("locked")
-                } else {
-                    Ok("done")
-                }
-            }
-        })
-        .await;
-        assert_eq!(out, Ok("done"));
-        assert_eq!(
-            calls.load(Ordering::SeqCst),
-            4,
-            "three transient failures, then success on the fourth try"
-        );
-    }
-
-    /// A file that stays locked for every attempt must surface the *last*
-    /// error (not be silently swallowed) after exactly `attempts` tries — the
-    /// property that lets a genuinely stuck binary still fail the install
-    /// loudly instead of hanging or pretending to succeed.
-    #[tokio::test]
-    async fn retry_transient_surfaces_error_after_exhausting_attempts() {
-        use std::sync::atomic::{AtomicU32, Ordering};
-        let calls = AtomicU32::new(0);
-        let out: Result<(), &str> = retry_transient(4, Duration::ZERO, || {
-            calls.fetch_add(1, Ordering::SeqCst);
-            async { Err("still locked") }
-        })
-        .await;
-        assert_eq!(out, Err("still locked"));
-        assert_eq!(
-            calls.load(Ordering::SeqCst),
-            4,
-            "exactly `attempts` calls are made — no more, no fewer"
-        );
-    }
-
-    /// `attempts == 1` means "try once, no retry" — the degenerate bound must
-    /// not off-by-one into zero attempts or an extra one.
-    #[tokio::test]
-    async fn retry_transient_with_single_attempt_does_not_retry() {
-        use std::sync::atomic::{AtomicU32, Ordering};
-        let calls = AtomicU32::new(0);
-        let out: Result<(), &str> = retry_transient(1, Duration::ZERO, || {
-            calls.fetch_add(1, Ordering::SeqCst);
-            async { Err("locked") }
-        })
-        .await;
-        assert_eq!(out, Err("locked"));
-        assert_eq!(calls.load(Ordering::SeqCst), 1);
-    }
 
     /// `rename_with_retry` moves the file and clears the source on the happy
     /// path — the same contract as a bare `rename`, so wiring it into
